dashboard/backend/pyproject.toml
[project]
name = "notebook-backend"
version = "0.1.0"
description = "Backend for new jupyter notebook"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3",
    "fastapi",
    "ipykernel",
    "jupyter-client",
    "sh",
    "uvicorn",
    "websockets",
    "supabase",
    "python-dotenv>=1.0.1",
    "apscheduler>=3.11.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary==2.9.10",
    "requests>=2.32.3",
    "resend>=2.5.1",
    "docker",
    "pyjwt>=2.8.0",
]
